backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1
PyPDF2==3.0.1
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0

# LLM providers
openai>=1.0.0
google-generativeai>=0.7.0
groq>=0.9.0

# Minimized OCR stack (avoid heavy torch/scikit-image)
pymupdf==1.24.9
# opencv/scikit-image/torch removed to reduce size

# Optional vector clients (commented to minimize footprint)
# sentence-transformers==2.7.0
# torch==2.3.1
# pinecone-client==3.2.2